chore: move class to a new file
