+++ libs/superagent/app/api/workflow_configs/processors/agent_processor.py
@@ -1,81 +1,10 @@
 from itertools import zip_longest
 
 from app.api.workflow_configs.api.api_manager import ApiManager
+from app.api.workflow_configs.data_transformer import DataTransformer
 from app.api.workflow_configs.processors.processor import Processor
+from app.api.workflow_configs.validator import SAMLValidator
 from app.utils.helpers import compare_dicts, get_first_non_null_key
-
-from ..data_transformer import DataTransformer
-
-
-class RepeatedNameError(Exception):
-    pass
-
-
-class Validator:
-    def __init__(self, config):
-        self.config = config
-
-    def validate(self):
-        self.validate_assistant_names()
-        self.validate_tool_names()
-        self.validate_superrag_names()
-
-    def validate_assistant_names(self):
-        assistants = self.config.get("workflows", [])
-        assistant_names = []
-
-        for new_assistant in assistants:
-            assistant_type = get_first_non_null_key(new_assistant)
-            assistant_name = new_assistant.get(assistant_type).get("name")
-
-            if assistant_name in assistant_names:
-                raise RepeatedNameError(
-                    f"Assistant name '{assistant_name}' is repeated in the SAML,"
-                    f"please use unique names for each assistant."
-                )
-            assistant_names.append(assistant_name)
-
-    def validate_tool_names(self):
-        assistants = self.config.get("workflows", [])
-
-        for new_assistant in assistants:
-            assistant_type = get_first_non_null_key(new_assistant)
-            assistant = new_assistant.get(assistant_type)
-
-            tools = assistant.get("tools", [])
-            tool_names = []
-
-            for tool in tools:
-                tool_type = get_first_non_null_key(tool)
-                tool_name = tool.get(tool_type).get("name")
-
-                if tool_name in tool_names:
-                    raise RepeatedNameError(
-                        f"Tool name '{tool_name}' is repeated in the SAML,"
-                        f"please use unique names for each tool."
-                    )
-                tool_names.append(tool_name)
-
-    def validate_superrag_names(self):
-        assistants = self.config.get("workflows", [])
-
-        for new_assistant in assistants:
-            assistant_type = get_first_non_null_key(new_assistant)
-            assistant = new_assistant.get(assistant_type)
-
-            superrag = assistant.get("superrag") or []
-            superrag_names = []
-
-            for superrag_data in superrag:
-                node_type = get_first_non_null_key(superrag_data)
-                superrag_name = superrag_data.get(node_type).get("name")
-
-                if superrag_name in superrag_names:
-                    raise RepeatedNameError(
-                        f"Superrag name '{superrag_name}' is repeated in the SAML,"
-                        f"please use unique names for each superrag."
-                    )
-                superrag_names.append(superrag_name)
 
 
 class AgentProcessor:
@@ -202,7 +131,7 @@ class AgentProcessor:
         return new_agent
 
     async def process_assistants(self, old_config, new_config):
-        validator = Validator(new_config)
+        validator = SAMLValidator(new_config)
         validator.validate()
 
         old_assistants = old_config.get("workflows", [])
